# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RemessasSteam API",
        "description": "API para gestão de projetos, parceiros e divisão de receitas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Listar projetos",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de projetos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Criar projeto",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "created_by",
                                    "title"
                                ],
                                "properties": {
                                    "created_by": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Projeto X"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "date_release": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "finished": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "steam_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "capsule": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projeto criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projects/{project}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Detalhar projeto",
                "operationId": "1352e06e55f6ab10d3702509cd385918",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projeto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Atualizar projeto",
                "operationId": "a93b65e3037bfd5ea5cdc1b3dd11ba53",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "created_by",
                                    "title"
                                ],
                                "properties": {
                                    "created_by": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Projeto X"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "date_release": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "finished": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "steam_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "capsule": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projeto atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Remover projeto",
                "operationId": "63c289a907902132dd95dd7ecb50143f",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Projeto removido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projects/{project}/partners": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Listar parceiros do projeto",
                "operationId": "4e19933b1a5803cafb6b843df504cc96",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parceiros do projeto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Sincronizar parceiros do projeto",
                "operationId": "220abba052b76978823199f40db7f174",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "partners"
                                ],
                                "properties": {
                                    "partners": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "partner_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "share_percent": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "valid_from": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "nullable": true
                                                },
                                                "valid_until": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "sync",
                                            "attach",
                                            "detach"
                                        ],
                                        "example": "sync"
                                    },
                                    "detach_missing": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parceiros sincronizados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}